Controller Complete

+++ src/main/java/com/vbt/factoryManagement/repository/AuthorityRepository.java
@@ -3,5 +3,5 @@ package com.vbt.factoryManagement.repository;
 import com.vbt.factoryManagement.entities.AuthorityList;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthorityRepository extends JpaRepository<AuthorityList,Integer> {
+public interface AuthorityRepository extends JpaRepository<AuthorityList,Long> {
 }
